+++ docs/src/changelog.md
@@ -7,6 +7,24 @@ CurrentModule = LibSSH
 This documents notable changes in LibSSH.jl. The format is based on [Keep a
 Changelog](https://keepachangelog.com).
 
+## [v0.7.0] - 2024-10-25
+
+### Added
+
+- [`Demo.DemoServer`](@ref) now supports passing `allow_auth_none=true` to allow
+  easily setting up passwordless authentication ([#28]).
+
+### Fixed
+
+- Previously the [`Demo.DemoServer`](@ref)'s command execution implementation
+  would only send the command output after it had finished. Now the output gets
+  sent as soon as it's printed by the command ([#28]).
+
+### Changed
+
+- **Breaking**: [`set_channel_callbacks()`](@ref) will remove any existing
+  callbacks ([#28]).
+
 ## [v0.6.1] - 2024-10-20
 
 ### Added